✨ chore: refactor logging setup across modules
Replace custom logging setup with a centralized logger 
configuration. This change improves logging consistency 
and maintainability by utilizing a shared logger setup 
function. Unused debug logging statements are removed 
to clean up the code.
