--- src/utils/logger_config.py
@@ -1,1 +1,0 @@
-
